+++ src/SOLOMotorControllersCanopen.h
@@ -5,8 +5,8 @@
  * @brief   This file contains all the base functions prototypes for the CANopen
  *          Availability: https://github.com/Solo-FL/SOLO-motor-controllers-ARDUINO-library
  *
- * @date    Date: 2024
- * @version 5.4.0
+ * @date    Date: 2025
+ * @version 5.5.0
  * *******************************************************************************
  * @attention
  * Copyright: (c) 2021-present, SOLO motor controllers project
@@ -159,7 +159,7 @@ public:
   /** @addtogroup CANOpen_PDO_Functions CANOpen Functions for Work with PDO Objects
    * @{
    */
-  #if defined(ARDUINO_PORTENTA_C33) || defined(ARDUINO_UNOWIFIR4) || defined(ARDUINO_MINIMA)
+  #if defined(ARDUINO_PORTENTA_C33) || defined(ARDUINO_UNOWIFIR4) || defined(ARDUINO_AVR_MEGA2560) || defined(ARDUINO_PORTENTA_H7_M7)
     virtual bool SetPdoParameterConfig(PdoParameterConfig config, int &error);
     virtual bool SendPdoSync(int &error);
 
@@ -178,7 +178,7 @@ public:
     virtual float GetPdoBoardTemperature(int &error);
     virtual PdoParameterConfig GetPdoParameterConfig(PdoParameterName parameterName, int &error);
     virtual bool UpdatePdoParameterCobIdByPdoParameterName();
-  #endif // ARDUINO_PORTENTA_C33 ARDUINO_UNOWIFIR4 ARDUINO_MINIMA
+  #endif // ARDUINO_PORTENTA_C33 ARDUINO_UNOWIFIR4 ARDUINO_MINIMA ARDUINO_PORTENTA_H7_M7
 
   /**
    * @}